+++ tests/OsType/ValueBuilders/sw_vers-examples/osx-10.11.2.php
@@ -1,0 +1,9 @@
+#!/usr/bin/env php
+<?php
+
+echo <<<EOS
+ProductName:	Mac OS X
+ProductVersion:	10.11.2
+BuildVersion:	15C50
+
+EOS;